Split tests into individual files
